# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sports prediction engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /snapshots:
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: List game snapshots
      parameters:
        - name: date
          in: query
          schema:
            type: string
            description: YYYY-MM-DD
        - name: league
          in: query
          schema:
            $ref: "#/components/schemas/League"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/GameStatus"
      responses:
        "200":
          description: List of snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSnapshot"
  /snapshots/generate:
    post:
      operationId: generateSnapshots
      tags: [snapshots]
      summary: Generate game snapshots for a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSnapshotsRequest"
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResult"
  /snapshots/finalize:
    post:
      operationId: finalizeCloses
      tags: [snapshots]
      summary: Finalize closing lines for a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeDateRequest"
      responses:
        "200":
          description: Finalization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResult"
...
components:
  schemas:
...
    League:
      type: string
      enum: [nba, ncaam, nhl]
...
    GameStatus:
      type: string
      enum: [scheduled, in_progress, final, cancelled]
...
    GameSnapshot:
      type: object
      properties:
        id:
          type: integer
        gameKey:
          type: string
        league:
          $ref: "#/components/schemas/League"
        eventStart:
          type: string
          format: date-time
        homeTeam:
          type: string
        awayTeam:
          type: string
        homePublishMl:
          type: number
        awayPublishMl:
          type: number
        homeCloseMl:
          type: number
          nullable: true
        awayCloseMl:
          type: number
          nullable: true
        publishSpread:
          type: number
          nullable: true
        publishSpreadLine:
          type: number
          nullable: true
        closeSpread:
          type: number
          nullable: true
        closeSpreadLine:
          type: number
          nullable: true
        publishTotal:
          type: number
          nullable: true
        publishOverLine:
          type: number
          nullable: true
        publishUnderLine:
          type: number
          nullable: true
        closeTotal:
          type: number
          nullable: true
        closeOverLine:
          type: number
          nullable: true
        closeUnderLine:
          type: number
          nullable: true
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        status:
          $ref: "#/components/schemas/GameStatus"
        snapshotDate:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - gameKey
        - league
        - eventStart
        - homeTeam
        - awayTeam
        - homePublishMl
        - awayPublishMl
        - status
        - snapshotDate
        - createdAt
...
    GenerateSnapshotsRequest:
      type: object
      properties:
        date:
          type: string
          description: YYYY-MM-DD
        leagues:
          type: array
          items:
            $ref: "#/components/schemas/League"
          default: [nba, nhl, ncaam]
        games:
          type: array
          items:
            $ref: "#/components/schemas/GameSnapshotInput"
      required:
        - date
    GameSnapshotInput:
      type: object
      properties:
        gameKey:
          type: string
        league:
          $ref: "#/components/schemas/League"
        eventStart:
          type: string
          format: date-time
        homeTeam:
          type: string
        awayTeam:
          type: string
        homePublishMl:
          type: number
        awayPublishMl:
          type: number
        publishSpread:
          type: number
          nullable: true
        publishSpreadLine:
          type: number
          nullable: true
        publishTotal:
          type: number
          nullable: true
        publishOverLine:
          type: number
          nullable: true
        publishUnderLine:
          type: number
          nullable: true
      required:
        - gameKey
        - league
        - eventStart
        - homeTeam
        - awayTeam
        - homePublishMl
        - awayPublishMl
    FinalizeDateRequest:
      type: object
      properties:
        date:
          type: string
          description: YYYY-MM-DD
      required:
        - date
...
    PipelineResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        count:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - success
        - message
        - count